refactor(web): split quill checklist logic into separate files
